Added file upload
This is done in the upload_file.c file and you can specify the
pathname with the -F flag. The image description is the filename
whether you like it or not. So a TODO is to make a custom description

And improved the usage() function

Updated usage()

removed description from main